third_party/blink/web_tests/external/wpt/html/cross-origin/usecredentials.tentative-expected.txt

This is a testharness.js-based test.
FAIL Top-level with use-credentials policy: navigating a frame to a null policy should fail. assert_equals: expected null but got Document node with 1 child
FAIL Top-level with use-credentials policy: navigating a frame from a use-credentials policy to a null policy should fail assert_equals: Navigation to null policy should fail expected null but got Document node with 1 child
FAIL Top-level with use-credentials policy: navigating a frame from an anonymous policy to a null policy should fail. assert_equals: Navigation to null policy should fail expected null but got Document node with 1 child
Harness: the test ran to completion.
